:recycle: Refactor: updateKeywordCount 메소드 저장 위치 변경 #32
