+++ src/main/java/site/moamoa/backend/service/component/command/keyword/KeywordCommandService.java
@@ -5,4 +5,6 @@ import site.moamoa.backend.web.dto.response.KeywordResponseDTO;
 public interface KeywordCommandService {
     //개인 최근 검색어 중 삭제
     KeywordResponseDTO.DeleteKeywordResult deleteRecentKeyword(Long memberId, String keyword);
+
+    void updateKeywordCount(Long memberId, String keyword);
 }
